+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -8,6 +8,7 @@ import woowacourse.auth.support.JwtTokenProvider;
 import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.domain.Customer;
 import woowacourse.shoppingcart.dto.LoginCustomer;
+import woowacourse.shoppingcart.exception.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.InvalidCustomerLoginException;
 import woowacourse.shoppingcart.exception.InvalidTokenException;
 import woowacourse.shoppingcart.util.HashTool;
@@ -26,13 +27,15 @@ public class AuthService {
 
     @Transactional
     public TokenResponse createToken(TokenRequest tokenRequest) {
-        if (!customerDao.checkValidLogin(tokenRequest.getLoginId(), HashTool.hashing(tokenRequest.getPassword()))) {
+        try {
+            Customer customer = customerDao.findByLoginId(tokenRequest.getLoginId());
+            checkPassword(customer, tokenRequest.getPassword());
+
+            String token = jwtTokenProvider.createToken(tokenRequest.getLoginId());
+            return new TokenResponse(token, customer.getUsername());
+        } catch (InvalidCustomerException | IllegalArgumentException e) {
             throw new InvalidCustomerLoginException();
         }
-
-        String token = jwtTokenProvider.createToken(tokenRequest.getLoginId());
-        Customer customer = customerDao.findByLoginId(tokenRequest.getLoginId());
-        return new TokenResponse(token, customer.getUsername());
     }
 
     public LoginCustomer findCustomerByToken(String token) {
@@ -42,5 +45,13 @@ public class AuthService {
         String payload = jwtTokenProvider.getPayload(token);
 
         return new LoginCustomer(customerDao.findByLoginId(payload));
+    }
+
+    public void checkPassword(Customer customer, String password) {
+        String hashedPassword = HashTool.hashing(password);
+
+        if (!customer.isSamePassword(hashedPassword)) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
     }
 }
